added realtime updation of messages in channel
